Separate generation of mod.rs and widgets
