+++ src/codegen/widgets.rs
@@ -1,17 +1,11 @@
-use std::io::Write;
 use std::path::Path;
 
 use analysis;
 use env::Env;
 use file_saver::*;
 use nameutil::*;
-use super::general;
 
-pub fn generate(env: &Env) {
-    let root_path = env.config.target_path.join("src").join("auto");
-
-    let mut mod_rs: Vec<String> = Vec::new();
-    let mut traits: Vec<String> = Vec::new();
+pub fn generate(env: &Env, root_path: &Path, mod_rs: &mut Vec<String>, traits: &mut Vec<String>) {
 
     for obj in env.config.objects.values() {
         if !obj.status.need_generate() {
@@ -28,21 +22,6 @@ pub fn generate(env: &Env) {
             |w| super::widget::generate(w, env, &class_analysis));
 
         let mod_name = module_name(split_namespace_name(&class_analysis.full_name).1);
-        super::widget::generate_reexports(env, &class_analysis, &mod_name, &mut mod_rs,
-            &mut traits);
+        super::widget::generate_reexports(env, &class_analysis, &mod_name, mod_rs, traits);
     }
-
-    generate_mod_rs(env, &root_path, mod_rs, traits);
-}
-
-fn generate_mod_rs(env: &Env, root_path: &Path, mod_rs: Vec<String>, traits: Vec<String>) {
-    let path = root_path.join("mod.rs");
-    save_to_file(path, env.config.make_backup, |w| {
-        try!(general::start_comments(w, &env.config));
-        try!(general::write_vec(w, &mod_rs));
-        try!(writeln!(w, ""));
-        try!(writeln!(w, "pub mod traits {{"));
-        try!(general::write_vec(w, &traits));
-        writeln!(w, "}}")
-    });
 }
